feat : split ViewOffsetKey and BackgroundBlur component
